feat(day4): removed answers from previous days and did day4 with a custom class
The most logical data structure for the board allowed for ease of storing the umbers and specific
helpful data regarding them was a custom class. So I wrote that as board.py. This made the actual
ask for both parts one and two VERY easy. part2 did require the addition of the winninPosition
function for the board, but not hard to do.